requirements.txt
behave==1.2.6
appium-python-client==3.1.1
selenium==4.12.0
